+++ deployments/changes/postgres/20151222_add_checksum.sql
@@ -1,0 +1,2 @@
+alter table deployment_tracker add column checksum varchar(32);
+create index on deployment_tracker (checksum);