+++ functions/usage.sh
@@ -5,6 +5,7 @@ function usage() {
                                     (default: postgres)
   -c|--confirm                    No arguments, automatically confirms choices 
                                     without user interaction
+  -C|--checksum-off               No arguments, turns off checksum logging
   -d|--database [name]            Database name to work with (will attempt to 
                                     auto-discover if -f is used, but can be 
                                     overwritten by using this variable)